oauth/auth/auth_service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "0.1.0"
description = ""
authors = ["Muhammad Talha <talhajoiyamuhammad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^42.0.7"
types-python-jose = "^3.3.4.20240106"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
